+++ dbt/adapters/databricks/impl.py
@@ -69,12 +69,10 @@ from dbt.adapters.spark.impl import DESCRIBE_TABLE_EXTENDED_MACRO_NAME
 from dbt.adapters.spark.impl import GET_COLUMNS_IN_RELATION_RAW_MACRO_NAME
 from dbt.adapters.spark.impl import KEY_TABLE_OWNER
 from dbt.adapters.spark.impl import KEY_TABLE_STATISTICS
-from dbt.adapters.spark.impl import LIST_RELATIONS_MACRO_NAME
 from dbt.adapters.spark.impl import LIST_SCHEMAS_MACRO_NAME
 from dbt.adapters.spark.impl import SparkAdapter
 from dbt.adapters.spark.impl import TABLE_OR_VIEW_NOT_FOUND_MESSAGES
 from dbt_common.clients.agate_helper import DEFAULT_TYPE_TESTER
-from dbt_common.clients.agate_helper import empty_table
 from dbt_common.exceptions import DbtRuntimeError
 from dbt_common.utils import executor
 
@@ -214,9 +212,8 @@ class DatabricksAdapter(SparkAdapter):
     def list_relations_without_caching(  # type: ignore[override]
         self, schema_relation: DatabricksRelation
     ) -> List[DatabricksRelation]:
-        kwargs = {"schema_relation": schema_relation}
         try:
-            results = self.execute_macro(LIST_RELATIONS_MACRO_NAME, kwargs=kwargs)
+            results = self.get_relations_without_caching(schema_relation)
         except DbtRuntimeError as e:
             errmsg = getattr(e, "msg", "")
             if check_not_found_error(errmsg):
@@ -228,13 +225,16 @@ class DatabricksAdapter(SparkAdapter):
 
         return [
             self.Relation.create(
-                database=database,
-                schema=schema,
+                database=schema_relation.database,
+                schema=schema_relation.schema,
                 identifier=name,
                 type=self.Relation.get_relation_type(kind),
+                comment=comment if comment != "" else None,
+                file_format=file_format,
+                owner=owner,
             )
-            for database, schema, name, kind in results.select(  # type: ignore[attr-defined]
-                ["database_name", "schema_name", "name", "kind"]
+            for name, comment, kind, file_format, owner in results.select(
+                ["name", "comment", "kind", "file_format", "owner"]
             )
         ]
 
@@ -277,11 +277,24 @@ class DatabricksAdapter(SparkAdapter):
 
         return relations
 
-    @available.parse(lambda *a, **k: empty_table())
     def get_relations_without_caching(self, relation: DatabricksRelation) -> Table:
+        if relation.is_hive_metastore():
+            return self._get_hive_relations(relation)
+        return self._get_uc_relations(relation)
+
+    def _get_uc_relations(self, relation: DatabricksRelation) -> Table:
+        kwargs = {"relation": relation}
+        tables = self.execute_macro("get_uc_tables", kwargs=kwargs)
+        return Table(
+            tables,
+            column_names=["name", "comment", "kind", "file_format", "owner"],
+            column_types=[Text(), Text(), Text(), Text(), Text()],
+        )
+
+    def _get_hive_relations(self, relation: DatabricksRelation) -> Table:
         kwargs = {"relation": relation}
 
-        new_rows: List[Tuple[Optional[str], str, str, str]]
+        new_rows: List[Tuple[str, Optional[str]]]
         if all([relation.database, relation.schema]):
             tables = self.connections.list_tables(
                 database=relation.database, schema=relation.schema  # type: ignore[arg-type]
@@ -291,129 +304,29 @@ class DatabricksAdapter(SparkAdapter):
             for row in tables:
                 # list_tables returns TABLE_TYPE as view for both materialized views and for
                 # streaming tables.  Set type to "" in this case and it will be resolved below.
-                type = row["TABLE_TYPE"].lower() if row["TABLE_TYPE"].lower() != "view" else ""
-                row = (row["TABLE_CAT"], row["TABLE_SCHEM"], row["TABLE_NAME"], type)
+                type = row["TABLE_TYPE"].lower() if row["TABLE_TYPE"].lower() != "view" else None
+                row = (row["TABLE_NAME"], type)
                 new_rows.append(row)
 
         else:
             tables = self.execute_macro(SHOW_TABLES_MACRO_NAME, kwargs=kwargs)
-            new_rows = [
-                (relation.database, row["database"], row["tableName"], "") for row in tables
-            ]
+            new_rows = [(row["tableName"], None) for row in tables]
 
         # if there are any table types to be resolved
-        if any(not row[3] for row in new_rows):
-            if is_hive_metastore(relation.database):
-                new_rows = self._get_hive_types(relation, new_rows)
-            else:
-                new_rows = self._get_uc_types(relation, new_rows)
+        if any(not row[1] for row in new_rows):
+            with self._catalog(relation.database):
+                views = self.execute_macro(SHOW_VIEWS_MACRO_NAME, kwargs=kwargs)
+                view_names = set(views.columns["viewName"].values())  # type: ignore[attr-defined]
+                new_rows = [
+                    (row[0], str(RelationType.View if row[0] in view_names else RelationType.Table))
+                    for row in new_rows
+                ]
 
         return Table(
-            new_rows,
-            column_names=["database_name", "schema_name", "name", "kind"],
-            column_types=[Text(), Text(), Text(), Text()],
+            [(row[0], None, row[1], None, None) for row in new_rows],
+            column_names=["name", "comment", "kind", "file_format", "owner"],
+            column_types=[Text(), Text(), Text(), Text(), Text()],
         )
-
-    def _get_hive_types(
-        self,
-        relation: DatabricksRelation,
-        new_rows: List[Tuple[Optional[str], str, str, str]],
-    ) -> List[Tuple[Optional[str], str, str, str]]:
-        kwargs = {"relation": relation}
-
-        with self._catalog(relation.database):
-            views = self.execute_macro(SHOW_VIEWS_MACRO_NAME, kwargs=kwargs)
-
-            view_names = set(views.columns["viewName"].values())  # type: ignore[attr-defined]
-            return [
-                (
-                    row[0],
-                    row[1],
-                    row[2],
-                    str(RelationType.View if row[2] in view_names else RelationType.Table),
-                )
-                for row in new_rows
-            ]
-
-    def _get_uc_types(
-        self,
-        relation: DatabricksRelation,
-        new_rows: List[Tuple[Optional[str], str, str, str]],
-    ) -> List[Tuple[Optional[str], str, str, str]]:
-        kwargs = {"relation": relation}
-
-        # Get view names and create a dictionary of view name to materialization
-        relation_all_tables = self.Relation.create(
-            database=relation.database, schema=relation.schema, identifier="*"
-        )
-
-        with self._catalog(relation.database):
-            views = self.execute_macro(SHOW_VIEWS_MACRO_NAME, kwargs=kwargs)
-            tables = self.execute_macro(
-                SHOW_TABLE_EXTENDED_MACRO_NAME,
-                kwargs={"schema_relation": relation_all_tables},
-            )
-        view_names: Dict[str, bool] = {
-            view["viewName"]: view.get("isMaterialized", False) for view in views
-        }
-        table_names: Dict[str, bool] = {
-            table["tableName"]: (self._parse_type(table["information"]) == "STREAMING_TABLE")
-            for table in tables
-        }
-
-        # create a new collection of rows with the correct table types
-        new_rows = [
-            (
-                row[0],
-                row[1],
-                row[2],
-                str(
-                    row[3]
-                    if row[3]
-                    else self._type_from_names(row[0], row[2], view_names, table_names)
-                ),
-            )
-            for row in new_rows
-        ]
-
-        return new_rows
-
-    def _parse_type(self, information: str) -> str:
-        type_entry = [
-            entry.split(":")[1].strip()
-            for entry in information.split("\n")
-            if entry.startswith("Type:")
-        ]
-
-        return type_entry[0] if type_entry else ""
-
-    def _type_from_names(
-        self,
-        database: Optional[str],
-        name: str,
-        view_names: Dict[str, bool],
-        table_names: Dict[str, bool],
-    ) -> DatabricksRelationType:
-        if name in view_names:
-            # it is either a view or a materialized view
-            return (
-                DatabricksRelationType.MaterializedView
-                if view_names[name]
-                else DatabricksRelationType.View
-            )
-        elif is_hive_metastore(database):
-            return DatabricksRelationType.Table
-        elif name in table_names:
-            # it is either a table or a streaming table
-            return (
-                DatabricksRelationType.StreamingTable
-                if table_names[name]
-                else DatabricksRelationType.Table
-            )
-        else:
-            raise DbtRuntimeError(
-                f"Unexpected relation type discovered: Database:{database}, Relation:{name}"
-            )
 
     def get_relation(
         self,
@@ -467,7 +380,30 @@ class DatabricksAdapter(SparkAdapter):
     def get_columns_in_relation(  # type: ignore[override]
         self, relation: DatabricksRelation
     ) -> List[DatabricksColumn]:
-        return self._get_updated_relation(relation)[1]
+        try:
+            rows: List[Row] = list(
+                self.execute_macro(GET_COLUMNS_COMMENTS_MACRO_NAME, kwargs={"relation": relation})
+            )
+            columns = []
+            for row in rows:
+                if row["col_name"].startswith("#"):
+                    break
+                columns.append(
+                    DatabricksColumn(
+                        column=row["col_name"], dtype=row["data_type"], comment=row["comment"]
+                    )
+                )
+        except DbtRuntimeError as e:
+            # spark would throw error when table doesn't exist, where other
+            # CDW would just return and empty list, normalizing the behavior here
+            errmsg = getattr(e, "msg", "")
+            found_msgs = (msg in errmsg for msg in TABLE_OR_VIEW_NOT_FOUND_MESSAGES)
+            if any(found_msgs):
+                columns = []
+            else:
+                raise e
+
+        return columns
 
     def _get_updated_relation(
         self, relation: DatabricksRelation
@@ -634,7 +570,6 @@ class DatabricksAdapter(SparkAdapter):
                 quote_policy=self.config.quoting,
             )
             for relation, information in self._list_relations_with_information(schema_relation):
-                logger.debug("Getting table schema for relation {}", str(relation))
                 columns.extend(self._get_columns_for_catalog(relation, information))
         return Table.from_object(columns, column_types=DEFAULT_TYPE_TESTER)